[#64347] move details form to `app/forms`
